refactor webpack plugins into classes
