+++ docs/guides/building-client-side-assets-with-webpack.md
@@ -2,7 +2,7 @@
 
 Page Kit apps use [Webpack] to build client-side assets. We provide a handful of Webpack plugins to configure it to our conventions. For consistency with other apps, you should use these plugins.
 
-[Webpack]: https://webpack.js.org/
+[webpack]: https://webpack.js.org/
 
 ## Setting up Webpack
 
@@ -15,15 +15,13 @@ npm install --save-dev webpack webpack-cli @financial-times/dotcom-build-base
 Create a `webpack.config.js` file in the root of your app. This is the default [Webpack configuration file]. You should add an `entry` option pointing at your client-side entry point (usually `client/main.js`), and `plugins` array initialising the Page Kit config plugin:
 
 ```js
-const base = require('@financial-times/dotcom-build-base')
+const { PageKitBase } = require('@financial-times/dotcom-build-base')
 
 module.exports = {
-	entry: {
-		scripts: 'client/main.js'
-	},
-	plugins: [
-		base.plugin()
-	]
+  entry: {
+    scripts: 'client/main.js',
+  },
+  plugins: [new PageKitBase()],
 }
 ```
 
@@ -31,23 +29,23 @@ Run `webpack --mode=development`. It should output the file `public/scripts.bund
 
 All Webpack configuration is allowed, but some options may conflict with Page Kit plugins, and it's better to have more things standard between Page Kit apps for developer familiarity and to improve performance and caching across all apps. So, try to avoid custom configuration, preferring the Page Kit plugins instead.
 
-[Webpack configuration file]: https://webpack.js.org/configuration/
+[webpack configuration file]: https://webpack.js.org/configuration/
 
 ## Using Javascript features
 
 If your app uses features like ECMAScript modules, JSX, or other things not supported in (all) browsers, you should use the [`@financial-times/dotcom-build-js`](../../packages/dotcom-build-js/README.md) plugin:
 
 ```diff
-const base = require('@financial-times/dotcom-build-base')
-+ const js = require('@financial-times/dotcom-build-js')
+const { PageKitBase } = require('@financial-times/dotcom-build-base')
++ const PageKitJs = require('@financial-times/dotcom-build-js')
 
 module.exports = {
 	entry: {
 		scripts: 'client/main.js'
 	},
 	plugins: [
-		base.plugin(),
-+		js.plugin()
+		new PageKitBase(),
++		new PageKitJs()
 	]
 }
 ```
@@ -57,8 +55,8 @@ module.exports = {
 Apps that use [Sass] should install the [`@financial-times/dotcom-build-sass`](../../packages/dotcom-build-sass/README.md) plugin, which allows you to specify a Sass file as a separate entry point:
 
 ```diff
-const base = require('@financial-times/dotcom-build-base')
-+ const sass = require('@financial-times/dotcom-build-sass')
+const { PageKitBase } = require('@financial-times/dotcom-build-base')
++ const { PageKitSass } = require('@financial-times/dotcom-build-sass')
 
 module.exports = {
 	entry: {
@@ -66,52 +64,52 @@ module.exports = {
 +		styles: 'client/main.scss',
 	},
 	plugins: [
-		base.plugin(),
-+		sass.plugin()
+		new PageKitBase(),
++		new PageKitSass()
 	]
 }
 ```
 
 Running `webpack --mode=development` again should output the file `public/styles.bundle.css`.
 
-[Sass]: https://sass-lang.com/
+[sass]: https://sass-lang.com/
 
 ## Using Bower packages
 
 Apps that need packages from [Bower] should install the [`@financial-times/dotcom-build-bower-resolve`](../../packages/dotcom-build-bower-resolve/README.md) plugin:
 
 ```diff
-const base = require('@financial-times/dotcom-build-base')
-+ const bowerResolve = require('@financial-times/dotcom-build-bower-resolve')
+const { PageKitBase } = require('@financial-times/dotcom-build-base')
++ const { PageKitBowerResolve } = require('@financial-times/dotcom-build-bower-resolve')
 
 module.exports = {
 	entry: {
 		scripts: 'client/main.js',
 	},
 	plugins: [
-		base.plugin(),
-+		bowerResolve.plugin()
+		new PageKitBase(),
++		new PageKitBowerResolve()
 	]
 }
 ```
 
-[Bower]: https://bower.io/
+[bower]: https://bower.io/
 
 ## Code splitting
 
 FT.com apps should use the [`@financial-times/dotcom-build-code-splitting`](../../packages/dotcom-build-code-splitting/README.md) plugin to take advantage of shared cached dependencies between apps.
 
 ```diff
-const base = require('@financial-times/dotcom-build-base')
-+ const codeSplitting = require('@financial-times/dotcom-build-code-splitting')
+const { PageKitBase } = require('@financial-times/dotcom-build-base')
++ const { PageKitCodeSplitting } = require('@financial-times/dotcom-build-code-splitting')
 
 module.exports = {
 	entry: {
 		scripts: 'client/main.js',
 	},
 	plugins: [
-		base.plugin(),
-+		codeSplitting.plugin()
+		new PageKitBase(),
++		new PageKitCodeSplitting()
 	]
 }
 ```
