+++ examples/building-sass-files/webpack.config.js
@@ -1,9 +1,9 @@
-const base = require('@financial-times/dotcom-build-base')
-const sass = require('@financial-times/dotcom-build-sass')
+const { PageKitBase } = require('@financial-times/dotcom-build-base')
+const { PageKitSass } = require('@financial-times/dotcom-build-sass')
 
 module.exports = {
   entry: {
     styles: './src/main.scss'
   },
-  plugins: [base.plugin(), sass.plugin()]
+  plugins: [new PageKitBase(), new PageKitSass()]
 }
