+++ packages/dotcom-build-bower-resolve/README.md
@@ -15,11 +15,11 @@ npm install --save-dev @financial-times/dotcom-build-bower-resolve
 After installing the package you must add it to the list of plugins in your project's `webpack.config.js` configuration file:
 
 ```diff
-+ const bowerResolve = require('@financial-times/dotcom-build-bower-resolve')
++ const { PageKitBowerResolve } = require('@financial-times/dotcom-build-bower-resolve')
 
 module.exports {
   "plugins": [
-+    bowerResolve.plugin()
++    new PageKitBowerResolve()
   ]
 }
 ```
