+++ packages/dotcom-build-bower-resolve/src/index.ts
@@ -1,4 +1,21 @@
-import { plugin } from './plugin'
+import BowerResolvePlugin from 'bower-resolve-webpack-plugin'
+import type webpack from 'webpack'
 
-export { plugin }
-export default plugin
+export class PageKitBowerResolve {
+  apply(compiler: webpack.Compiler) {
+    compiler.options.resolve = {
+      // This will find .js files if a bower.json specifies a list of "main" files
+      plugins: [new BowerResolvePlugin()],
+      // In which folders the resolver look for modules relative paths are
+      // looked up in every parent folder (like node_modules) absolute
+      // paths are looked up directly the order is respected
+      modules: ['bower_components', 'node_modules'],
+
+      // package description files
+      descriptionFiles: ['bower.json', 'package.json'],
+
+      // file names for directory resolution
+      mainFiles: ['index', 'main']
+    }
+  }
+}
