+++ packages/dotcom-build-code-splitting/README.md
@@ -15,11 +15,11 @@ npm install --save-dev @financial-times/dotcom-build-code-splitting
 After installing the package you must add it to the list of plugins in your project's `webpack.config.js` configuration file:
 
 ```diff
-+ const codeSplitting = require('@financial-times/dotcom-build-code-splitting')
++ const { PageKitCodeSplitting } = require('@financial-times/dotcom-build-code-splitting')
 
 module.export = {
   plugins: [
-+    codeSplitting.plugin()
++    new PageKitCodeSplitting()
   ]
 }
 ```
